refactored file-system for client files, also worked on refactoring index.js and App.jsx
