+++ client/Components/SearchBar.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom'
 import { useSelector, useDispatch } from 'react-redux';
-import { update_allergens, fetchProductsByAllergen, update_results } from '../productsSlice';
+import { update_allergens, fetchProductsByAllergen, update_results } from '../Reducers/productsSlice';
 
 const SearchBar = () => {
   // dispatch actions using RTK
