+++ client/index.html
@@ -4,7 +4,7 @@
     <head>
       <link rel="stylesheet" type="text/css" href="//fonts.googleapis.com/css?family=Lato" />
       <meta charset="UTF-8" />
-      <title>Search Skincare By Ingredient</title>
+      <title>SafeFace: Search Skincare By Ingredient</title>
     </head>
     <body>
       <!-- Hang React App Here -->
