+++ client/index.js
@@ -1,38 +1,19 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import { createRoot } from 'react-dom/client';
 import './styles.scss';
-import App from './Routes/App (Root).jsx';
-import store from './store.js';
-import Error from './Routes/Error.jsx';
-import SearchPage from "./Routes/SearchPage.jsx";
-import ResultsPage from "./Routes/ResultsPage.jsx";
+import App from './App.jsx';
+import store from './Reducers/store.js';
 
-// this sets the root of the DOM
-const domNode = document.getElementById('root');
-const root = createRoot(domNode);
+// Grabbing the element on 'index.html' with the ID of 'root' to render our 'App' component on
+const rootContainer = document.getElementById('root');
+const root = createRoot(rootContainer);
 
-// these are the possible routes react routers knows to use, including a general error page
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <App />,
-    errorElement: <Error />,
-  },
-  {
-    path: '/main',
-    element: <SearchPage />,
-  },
-  {
-    path: '/results',
-    element: <ResultsPage />,
-  },
-]);
-
-// this renders our pages based on the routes above, all within the redux store
+// Refactored index.js to just render the 'App' component, keeping the <Provider> wrapper to
+// allow access to the store for all child components
 root.render(
   <Provider store={store}>
-    <RouterProvider router={router} />
+    <App />
+    {/* <RouterProvider router={router} /> */}
   </Provider>
 );
