+++ client/styles.scss
@@ -1,5 +1,4 @@
 // these are sass color variables
-
 $primary_text: black;
 $secondary_text: rgb(92, 92, 92);
 $tertiary_text: darkgray;
@@ -34,6 +33,10 @@ $tertiary_text: darkgray;
   text-align: center;
 }
 
+.search-page h2 {
+  margin-bottom: 0;
+}
+
 .search-area {
   position: relative;
   top: 25%;
@@ -51,24 +54,25 @@ $tertiary_text: darkgray;
   border-bottom: 2px solid rgba(0, 0, 255, 0.499);
   width: 20%;
   text-align: center;
+  padding: 0.5rem;
+  border-radius: 0.5rem;
 }
 
 #search-button {
-  background-color: blue;
+  background-color: rgb(72, 72, 255);
   border: none;
-  border-radius: 10px;
+  border-radius: 5px;
   color: white;
   padding: 10px 32px;
-  margin: 20px 20px;
-  text-align: center;
-  text-decoration: none;
+  margin: 20px;
   display: inline-block;
   font-size: 16px;
+  cursor: pointer;
 }
 
 #search-button:hover {
-  box-shadow: 0 12px 16px 0 rgba(0, 0, 0, 0.24),
-    0 17px 50px 0 rgba(0, 0, 0, 0.19);
+  box-shadow: 
+    0 8px 10px 0 rgba(0, 0, 0, 0.24),
 }
 
 // Specific styling for Results Page and components
